Add init method prior registration task
For creating the symlink structure to a container application
a directory is needed which stores the container symlinks. This
commit adds the needed init method to create this directory prior
any registration tasks. All container symlinks will be stored
in /usr/share/flakes. This means an application link structure
for "app@container" would look like this:

    /usr/bin/app -> /usr/share/flakes/container -> /usr/bin/oci-pilot

A flake is also the name we invented to handle and package
OCI containers which gets managed by oci-pilot