Fix engine option handling
Runtime options were handled as hash values whereas the
hash key was used as option name and the hash value as
optional option value. This concept does not work for
options that can be specified multiple times because the
hash key must be unique or when options must be in order.
This commit changes the option handling such that the options
are provided as simple list values. Each value of the list
uses a space to separate the option name from its value.
The split by space happens only once and produces a tuple.
This Fixes #28