Add default network setup
Use initrd options rd.neednet=1 ip=dhcp to let systemd inside
of the initrd know how to setup the network. This design requires
the firecracker image to use an initrd and that initrd must use
systemd (typically created by dracut) in order to process the
above cmdline arguments. If this concept is not acceptable by
the users use case (e.g no initrd wanted) then the user needs
to provide its own network setup mechanics as part of the
firecracker rootfs image and also need to find a way to activate
it prior launching the application, e.g by using systemd instead
of sci as the init process.