Addded man page for sci guest utility
sci is used in combination with the firecracker engine
and allows to execute the registered application inside
of the guest VM. It also takes over responsiblity to
setup an eventual overlay as well as perform the shutdown
of the VM after the command execution. Further versions
of sci will also implement the vsoc communication channels
between host and guest. This commit adds a first version
of the man page to describe sci in more detail.