Refactor sci and support resume mode
The former implementation used the vsock channel to
read the command, call it in a child and send the output
back through the vsock channel. This works until the command
is not interactive. With this refactor I propose a different
way to handle the command execution.

start vsock listener on VM_PORT, wait for command(s) in a loop
A received command turns into an socat process calling
the command with an expected listener
Example:
    sudo socat UNIX-CONNECT:/run/sci_cmd_XXX.sock -
    CONNECT defaults::VM_PORT(52)
    --> send the command to call and quit

    sudo socat VSOCK-CONNECT:2:exec_port EXEC:cmd
    --> connects to the listener instance on the host (pilot)

The above procedure got also implemeted as
part of the firecracker-pilot resume code.
This Fixes #87