Keep comment format consistent
Maybe nitpick but I think consistency makes things easier.
I only adapted the comment block for this place in code
as it was mine before. There are other places which I did not touch
since I don't want to step on other peoples toes, but I recommend
that we decide for one comment format and stick to it everywhere