Firecracker sanity check
The firecracker startup procedure in resume mode needs to setup
the control channels prior launching the actual firecracker program
to allow synchronisation of guest vs. host data flow. If there is
no check that the firecracker binary exists we still perform all
the pre setup and waiting calls and the actual error condition is
only visible in debug mode. Therefore an early exit condition that
checks if there is a firecracker binary found on the host improves
the user experience. This Fixes #129