Move is_debug to common lib
Along with the change update podman and firecracker pilots
to make use of is_debug from the common space and print
debug messages through the debug! macro. The former double
formatting of the debug message broke the output display

In addition to the debug changes also a method do_trace()
was added to create trace reports. However, the method is
currently not in use as I'm still looking for a way to
create a useful trace report without spreading do_trace()
over the code