Use raw channel fallback if pty setup failed
In resume mode the firecracker pilot uses a vsock connection
for the host <-> guest data transfer. On the guest the command
is by default called in a new PTY. However if this is not possible
e.g because the guest does not support pty allocation there
should be a fallback allocating the standard channels directly.