Add support for force vsock on firecracker pilot
The vsock communication between guest and host was established
with the firecracker pilot only for resume type registrations.
For those it's required because you cannot mux one serial
terminal between parallel applications running at the same
time in the guest. So in resume mode each call has its own
socket and channels. However, there is no reason to not also
support that type of data flow between guest and host for one
time calls which fires up the VM, starts the application and
shuts down the VM. This commit adds support for forcing the
vsock communication also for the simple app registration.
The default stays on serial though.