Drop build-deb support
Building a debian package from a given oci container using the
debbuild tool was an extra helper command to turn a container
into a package. However, flake-pilot should not have this
responsibility and if it should provide it, it should be done
with the native tools from the respective distribution.