Handle incomplete container path
At present in order to load a container the full file name must be specified.
In an automated setup where the container is delivered by a source such as
an RPM package and then loaded by another process the exact filename may
not be known. We expect that the user provide the path to the tarball
including the beginning of the archive name. If multiple archives match the
highest file based on alpha-numeric sorting is used to load as a container.

Fixes #51