Better error handling for config_from_str()
If the provided source data is empty this is not an error
for YamlLoader and the empty data is just passed along as
a None value. The unconditional unwrap of a None value
however, let the application to panic into a stacktrace.
This commit catches the condition and provides a better
error message.