Improve error handling
Check for the presence of the main flake registration.
Even though this should never happen because the file
is the result of a successful flake-ctl registration we
cannot prevent users from working with it outside
flake-ctl. In addition explicitly match on potential
parse errors from aml_rust::YamlLoader instead of looking
at the error message as result from an Err unwrap.