Use pull policy set to: newer
if the image on the registry is newer than the one in the
local flake registry, make sure to fetch the latest version
automatically. Pull errors are suppressed if a local image
was found and we can't pull from the source location e.g an
image that was provided by a package and the blob was loaded
from the local storage.